Implementing all nest routes for Motorcycle usecases

+++ application/usecases/motorcycle/RemoveMotorcycle.ts
@@ -1,9 +1,16 @@
 import { MotorcycleRepository } from "../../repositories/MotorcycleRepository";
+import { VehicleNotFound } from "../../../domain/errors/VehicleNotFound";
 
 export class RemoveMotorcycle {
   constructor(private motorcycleRepository: MotorcycleRepository) {}
 
-  async execute(id: string): Promise<void> {
+  async execute(id: string): Promise<void | VehicleNotFound> {
+    const motorcycle = await this.motorcycleRepository.findById(id);
+
+    if (!motorcycle) {
+      return new VehicleNotFound();
+    }
+
     await this.motorcycleRepository.delete(id);
   }
 }
